Fix : store 이름 변경에 따른  file change
