--- stores/searchKeywordStore.ts
@@ -1,11 +1,0 @@
-import { create } from 'zustand';
-
-interface searchKeywordProps {
-  searchKeyword: string;
-  setSearchKeyword: (keyword: string) => void;
-}
-/**서버에 보낼 최종 키워드를 관리하는 store*/
-export const useSearchKeywordStore = create<searchKeywordProps>((set) => ({
-  searchKeyword: '',
-  setSearchKeyword: (keyword: string) => set({ searchKeyword: keyword }),
-}));
